MeetYou/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Django==2.2.18
django-ckeditor==6.0.0
django-guardian==2.3.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-notifications-hq==1.6.0
djangorestframework==3.12.2
iniconfig==1.1.1
jsonfield==3.1.0
packaging==20.9
pip==20.2.3
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
setuptools==49.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
